Clang-tidy fixes to the gui library
Mostly adding explicit to constructors, but also fixed a slicing issue
with pugi::xml_document. This is in preparation for adding <template>,
which is needed for the load game menu.

+++ test/gui/test_ui_element.hpp
@@ -23,7 +23,7 @@ class TestUiElement : public gui::UiElement {
   }
 
   int getArea() {
-    return mWidth * mHeight * mScale;
+    return static_cast<int>(mWidth * mHeight * mScale);
   }
 
   gui::TraitTypeId userTraitType(int index) const override {
